Add a TileField database field type
Add TileField, a wrapper over a charfield which deserializes into an
api.utils.Tile instance.

Add TileCollection, which represents a group of tiles and can be
serialized / deserialized as a string like a regular tile.

Add LazyTile and LazyTileCollection, which are wrappers around Tile and
TileCollection respectively. These wrappers allow us to defer
deserialization of their underlying object unless absolutely necessary
(which it isn't when we're simply serving this data to the client).